vms/txs/mempool: unify avm and platformvm mempool implementations
